Rename these folders so that the problem names match the ones found in Sean Prashad's "LeetCode Patterns."
Link: https://seanprashad.com/leetcode-patterns/
